+++ ChangeLog.markdown
@@ -1,9 +1,15 @@
 Changelog for Imbo
 ==================
 
-Imbo-2.x.x
+Imbo-2.1.1
 ----------
-__N/A__
+__2016-03-11__
+
+* #437: Fixed issue with metadata not being formatted correctly using the new Mongo adapters (Christer Edvartsen)
+
+Imbo-2.1.0
+----------
+__2016-03-10__
 
 * #436: Added new MongoDB adapters that use the mongodb extension (Christer Edvartsen)
 * #434: Custom models for group(s) and access rule(s) (Christer Edvartsen)